SAD806x 1.3b
Compared to 1.3a.

No real big changes and this is why I have not updated previous provided documentation.
- Many small issues were corrected.
- A 'Recent files' menu has been added, for main files.
Yes it was asked for that, from the beginning, but I did not want to use registry, ini files or something in the exe directory.
With the repository, things had changed, so... Generated in the exe directory, so be sure to have the rights.
- A 'Settings/Text Output...' menu has been added in the tools menus.
Same principle, than for 'Recent files', generated in the exe directory, so be sure to have the rights.
By default 'Activate Settings' is not checked, disassembly text output is the same than for 1.3a version.
When 'Activate Settings' is checked, disassembly text output is generated in a different way with default values (not so different).
Then you can update some values (width, labels) to match your expectations.
- Scale precision option is now added for all scale expressions.
It is visible in tool and in text output.
It is also mapped with Xdf format (Import/Export/Copy/Paste)
- Tool screen has a bit changed, it is wider, because of new fields.
Data output has a bit changed too, for performance and for being easier to read.
- A 'Mass Update' menu has been added in tools menu.
For now one option in it, to give a specific format for comments, on the first line.
Useful for Xdf import Export and to have the same format between definitions.